replaced read_csv with read.csv to address RStudio crashing
Multiple users reported crashing with RStudio when using read_csv in loops.

I addressed this by replacing read_csv with read.csv.

I also removed the chunk of text that was used to convert the 1 column dataframe read in with read_csv to 2 column dataframess This was originally included when the initial raw data was formatted in a different way than it is formatted now.